fix package details navigation
